BackEnd/ClickNetWebhook/requirements.txt
pycryptodome
requests
fastapi
sqlalchemy
pymysql
uvicorn
debugpy
httpx
cryptography
kafka-python
aiokafka
python-dotenv
asyncmy
uvicorn[standard]
slowapi
stripe
openai
twilio